requirements.txt
# requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pandas>=1.2.0
opencv-python>=4.5.0
albumentations>=1.0.0
flask>=2.0.0
pillow>=8.0.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
wandb>=0.12.0
tqdm>=4.60.0
pyyaml>=5.4.0